Implemented loading beers list using punkapi

+++ app/src/main/java/com/example/beer/rest/model/Amount__1.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 class Amount__1 {
     @SerializedName("value")
     @Expose
-    var value: Int? = null
+    var value: Double? = null
 
     @SerializedName("unit")
     @Expose
